09-hugging-face-with-image-audio-ai/requirements.txt
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
datasets
transformers
matplotlib
librosa
soundfile
setuptools
sounddevice
accelerate
optimum
pillow
opencv-python
timm
